Writing distance matrices to file
scipy.ndimage.distance_transform_edt(matrix) does this for us, but unfortunately it calculates the difference to the 0's. The matrix therefore needs to be inverted first. I did a few tests on the spherical pore in the solid, and it seems that the inverse is correct. I haven't found a way to visualize/test the distance transform just yet.

The script now writes the distance matrix to a .npy-file for each threshold value. It writes the files to a folder called Distmatrices (that is not automatically created, which is also an issue with Plots. Automating this should be done at some point).