Changed one stdv to the corrected version
i.e. Nt*M*(Nb-i) --> Nt*M*(Nb-i)-1 in costheta_totav_rms. I wasn't using that rms for anything other than plotting, so the change won't be visible. I was using the uncorrected standard deviation because there is an issue with dividing with zero for costheta_eachMandt_rms, where we have Nb-i pairs of bonds. So for Nb-(Nb-1) we can't use the corrected version there as the denominator would be 0.