Upload before major change
I want to try to wrap a part of my code into a function. Calling a function over and over again can be time-consuming, but this way I avoid copy-pasting. I will test the code against the old code and check what is faster. If the new code is unreasonably slow, I am going to put the processing in the loop and not in the function.