Implemented search part
The program is slow, but has not crashed yet. i should compare the runtime of this script with the time runtime of my old script.

Also, I need to verify the results from this code! See that the configurations after the voxellation and thresholding looks the same for both versions of the code (as I have already tested the old implementation).