Well I TRIED to make it faster
Used jit, but does not speed up (quite the opposite, but now I am running it at the same time as some heavy simulation, so that might be it).

Next on the list:
I should fix it so that the size of the box don't vary (i.e. the height in the z-direction is constant. Same number of voxels, the voxel length stays the same). A variable box size will affect the percolation.

All this is using NVT. NPT is more of a challenge. One that I may have to tackle at some point.