File relies on too large files
I will make a change where i only look at the free bead for most the time. Need to read in from two files: One with only the bead for many time steps, another for the entire system for ONE time step (to make an estimate of the extent of the chain).