Fixed shift of substrate
Thought about shifting with +xmin and +ymin, but forgot to actually do it. It is fixed now. Next up: Maybe check if the substrate bead and the polymer bead overlap, and if they do, not put the substrate bead there? (As of now the polymer bead and the substrate bead have the same LJ interactions, so the LJ term would double up)