Bead diffusion: Averages
Implemented averages in analysis of bead-in-brush and bulk bead simulations. Also included a conversion to SI units, which yielded a more sensible constant term in the line fit. Made some additional plots. Will go on to only look at sections of the data in order to find an appropriate diffusion constant. That might be hard.