New approach to brush diffusion, start
Aaaaaaand I should use the same equilibrated system for all my diffusion attempts. I should vary the initial position of the diffusing bead, not the velocity. This means I need to use a few more steps than I originally thought.

The in.-file is OK, but I am not done with makesystem_placebead yet. I need to make another in.-file that reads the data from placebead and uses that.