Placing bead in eq. syst. seems to work
I compared LAMMPS data.-files and my output file from makesystem_placebead.py, and it seems correct. I copy-pasted most of the file, but changed the number of atoms and added one line in Atoms and one in Velocities.

The velocity of the bead is set from the equipartition theorem (with f=3 as for a monatomic gas). I assumed the velocity components to be the same for simplicity. Another option is to draw random velocity components according to the normal distribution, but then I'd need to figure out the variance. Since this is just one bead it should not matter much as long as the velocity is in the vicinity in v from the equipartition theorem.