Writing 'metadata' to file
Actually just writing the time interval (in terms of indices) of the fit. That used to be in the file name, but I actually need a lot of different intervals, so it would be cumbersome to keep it there. I can probably stitch these meta-files together eventually when I compare D vs sigma or D vs d.