C++ voxellation should be pretty complete
Have not made a test-run since last time (ran a few other simulations and the computer froze), so the code might not be perfect.

I am not entirely sure how I should write to file. I have done something now, but I am not sure which ordering of xyz is neater. Right now I need to keep track of a lot of stuff to know the coordinates of each voxel. I might want to write the coordinates to file. I have set aside some file names for that, so writing the coordinates to file shouldn't be that hard. Can do that now.

The (output) filenames are pretty arbitrary. I need to consider what folder I'm going to write to, etc. Right now I need a new Python-script to read the results, because the one(s) I have now read the 3D matrix saved to file by np.save. The format is going to be completely different.

Oh, and I only use the basic "for every voxel find the smallest distance"-approach to voxellation here. I might want to used linked lists or the like in the future.