Fixed error in bulk and added a script
The new script cuts the walk after it has moved past some (absolute) coordinate of z. Meant to mimic the limits imposed in the brush diffusion for better comparison. (Because we only look at the bead inside the brush for brush diffusion, and hence we cut the walk when the bead has exited the brush. So as time goes on we select the trajectories that are slower in the z-direction).