Safety upload
I was too preoccupied to upload these when I made the changes. Somewhere in there is a crucial bug fix. I think maybe it is dividing by the number of walks for bulk, or maybe I did that earlier. I know I removed (0,0) from the plots of D(d), which is an improvement.

There might be some new files in here as well, or some trivial changes (i.e. changing a parameter, etc.). Better safe than sorry, though.