More flexibility on d vs phi + visualization
Have used r=0.5 in order to not have overlaps, now I have generalized to r=1 (but should still be wrong for d=1.5, should fix that).

Anders suggested I look at a 2D system instead of 3D in order to find the porosity. This is implemented in d_vs_phi_fromflat.py. plot_diskgrid is used to aid this. I think I could use the results from here in order to fix d=1.5 for 3D.