Script for playing around with the input current
Used a standard current in my former script, so to avoid confusion I made a new one for playing around with the current. Saves figures to folders named by the current (duration and amplitude) wth a file name that conveys the same information (since I'm often uploading to the web).