Improved plot name + bug fixes
As of now, I get no Ca-curves because the model With cell-id 496497595 does not have Ca-dynamics in the dendrites. Might be handy later on. I am considering skipping the loop over all cell models and instead give one cell-id at the top of the file (or at least before the cell loop). That loop is kind of broken anyways (a Windows issue, I seem to recall). Maybe I'll make that change if I include more cells or find a cell model that is even better.