Startet work on dend.prop-loop script
Since I'm only ALMOST toast.

There is plenty to do, though:
*Double check file names
*Write to file in a sensible manner
*Plot in a sensible manner
*Implement a bool for plotting (old plots only)
*Find averages and rms'es
*Implement a test to ensure that nseg!=0. Otherwise we have a problem (maybe I'll do that right now. A warning and break should suffice)

I think I'll skip the old NEURON-enumeration of the segments. That does not provide a lot of (if any?) extra information, so it is probably not worth the hassle. I have other scripts if I need to double check anything.